* Remove circles_computeTransfer * Don't serialize numbers as hex when returning a DatabaseQueryResult from rpc * PostgresDb: Read NpgsqlType.Numeric as BigInteger

+++ Circles.Index.Postgres/PostgresDb.cs
@@ -1,4 +1,5 @@
 using System.Data;
+using System.Numerics;
 using System.Text;
 using Circles.Index.Common;
 using Nethermind.Core.Crypto;
@@ -66,6 +67,12 @@ public class PostgresDb(string connectionString, IDatabaseSchema schema) : IData
                 }
                 else
                 {
+                    if (table.Value.Namespace.StartsWith("V_"))
+                    {
+                        // Dirty way to skip indexes and primary keys for views
+                        continue;
+                    }
+
                     primaryKeyDdl.AppendLine(
                         $"ALTER TABLE \"{table.Value.Namespace}_{table.Value.Table}\" ADD PRIMARY KEY (\"blockNumber\", \"transactionIndex\", \"logIndex\"{additionalKeyColumnsString});");
                 }
@@ -138,6 +145,12 @@ public class PostgresDb(string connectionString, IDatabaseSchema schema) : IData
 
         foreach (var column in indexedColumns)
         {
+            if (@event.Namespace.StartsWith("V_"))
+            {
+                // Dirty way to skip indexes and primary keys for views
+                continue;
+            }
+
             string indexName = $"idx_{@event.Namespace}_{@event.Table}_{column.Column}";
             ddlSql.AppendLine(
                 $"CREATE INDEX IF NOT EXISTS \"{indexName}\" ON \"{@event.Namespace}_{@event.Table}\" (\"{column.Column}\");");
@@ -263,20 +276,33 @@ public class PostgresDb(string connectionString, IDatabaseSchema schema) : IData
         }
 
         using var reader = command.ExecuteReader();
+
+        var resultSchema = reader.GetColumnSchema().ToArray();
         var columnNames = new string[reader.FieldCount];
         for (int i = 0; i < reader.FieldCount; i++)
         {
             columnNames[i] = reader.GetName(i);
         }
-        
+
         var resultRows = new List<object[]>();
         var row = new object[reader.FieldCount];
         while (reader.Read())
         {
-            reader.GetValues(row);
+            for (int i = 0; i < reader.FieldCount; i++)
+            {
+                if (resultSchema[i].NpgsqlDbType == NpgsqlDbType.Numeric)
+                {
+                    row[i] = reader.GetFieldValue<BigInteger>(i);
+                }
+                else
+                {
+                    row[i] = reader.GetValue(i);
+                }
+            }
+
             resultRows.Add(row);
         }
-        
+
         return new DatabaseQueryResult(columnNames, resultRows);
     }
 
@@ -295,6 +321,12 @@ public class PostgresDb(string connectionString, IDatabaseSchema schema) : IData
         {
             foreach (var table in Schema.Tables.Values)
             {
+                if (table.Namespace.StartsWith("V_"))
+                {
+                    // Dirty way to skip views
+                    continue;
+                }
+
                 await using var command = connection.CreateCommand();
                 command.CommandText =
                     $"DELETE FROM \"{table.Namespace}_{table.Table}\" WHERE \"{"blockNumber"}\" >= @reorgAt;";
